refactor : memberstudyroomRepositoryDSL 을 우한 디렉 구조 변경
