+++ src/main/java/com/linkode/api_server/repository/memberstudyroom/MemberstudyroomRepository.java
@@ -1,19 +1,15 @@
-package com.linkode.api_server.repository;
+package com.linkode.api_server.repository.memberstudyroom;
 
 import com.linkode.api_server.domain.base.BaseStatus;
 import com.linkode.api_server.domain.memberstudyroom.MemberRole;
 import com.linkode.api_server.domain.memberstudyroom.MemberStudyroom;
-import com.linkode.api_server.dto.studyroom.DetailStudyroomResponse;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
 @Repository
 public interface MemberstudyroomRepository extends JpaRepository<MemberStudyroom, Long> {
@@ -26,16 +22,6 @@ public interface MemberstudyroomRepository extends JpaRepository<MemberStudyroom
     int deleteMemberStudyroom(long studyroomId);
 
     Optional<MemberStudyroom> findByMember_MemberIdAndStudyroom_StudyroomIdAndStatus(Long memberId, Long studyroomId, BaseStatus status);
-    Optional<List<MemberStudyroom>> findByMember_MemberIdAndStatus(Long memberId, BaseStatus status);
-    Optional<List<MemberStudyroom>> findByStudyroom_StudyroomIdInAndStatus(Set<Long> studyroomIds, BaseStatus status);
-    @Modifying
-    @Query("UPDATE Studyroom s SET s.status = :status WHERE s.studyroomId IN :studyroomIds")
-    void updateStudyroomStatus(@Param("studyroomIds") Set<Long> studyroomIds, @Param("status") BaseStatus status);
-
-    @Modifying
-    @Query("UPDATE MemberStudyroom ms SET ms.status = :status WHERE ms IN :memberStudyrooms")
-    void updateMemberStudyroomStatus(@Param("memberStudyrooms") List<MemberStudyroom> memberStudyrooms, @Param("status") BaseStatus status);
-
     @Query("SELECT ms " +
             "FROM MemberStudyroom ms " +
             "JOIN FETCH ms.studyroom s " +
